SLFLF-3: Move CupertinoFormSection to list_section.dart
